BE/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
sqlalchemy[asyncio]
asyncpg
pyjwt
groq
